updateChecker: split out relevant logic from attachEarlyEndpoints
We'll need this in its own place so we can call it without going
through the install API. The install API can handle the web browser
response on its own.
